// Dashboard/src/Types/Tables.type.ts
export interface Questions {
    _id: string;
    categoryid: {
        name: string;
    };
    subcategoryid: {
        subcategory: string;
    };
    title: string;
    file: string;
}